#3 clean up
