+++ client/app/layout.tsx
@@ -5,7 +5,7 @@ import "./globals.css";
 const inter = Inter({ subsets: ["latin"] });
 
 export const metadata: Metadata = {
-  title: "Create Next App",
+  title: "Eduventure",
   description: "Generated by create next app",
 };
