--- server/db_helper_functions.py
@@ -1,3 +1,0 @@
-INSERT_USER = (
-    "INSERT INTO User(username, homeUniversity) VALUES (%s, %d)"
-)
